Moved tasks-navigation-accordion
